backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
google-generativeai
